requirements.txt
apache-beam[gcp]
pyiceberg[gcp,pyarrow]
pyarrow
google-auth
requests